Migrate codebase and unit tests to relocated and renamed types
