+++ tests/fixtures/reference_quiz_single_attempt_no_backups.py
@@ -13,11 +13,12 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from archiveworker.custom_types import JobArchiveRequest
+
+from archiveworker.api.worker import QuizArchiverRequest
 from tests.fixtures import reference_quiz_full
 
 ARCHIVE_API_REQUEST = {
-    "api_version": JobArchiveRequest.API_VERSION,
+    "api_version": QuizArchiverRequest.API_VERSION,
     "moodle_base_url": "http://localhost",
     "moodle_ws_url": "http://localhost/webservice/rest/server.php",
     "moodle_upload_url": "http://localhost/webservice/upload.php",
