+++ tests/test_demomode.py
@@ -22,7 +22,7 @@ import zipfile
 
 import pytest
 
-from archiveworker.custom_types import JobStatus
+from archiveworker.type import JobStatus
 from archiveworker.moodle_quiz_archive_worker import start_processing_thread
 from config import Config
 from .conftest import client, TestUtils
